% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraits.R
\docType{package}
\name{BayesTraitsAnalysis-package}
\alias{BayesTraitsAnalysis}
\alias{BayesTraitsAnalysis-package}
\title{BayesTraitsAnalysis}
\description{
Analysis of BayesTraits log files.
}
\details{
\tabular{ll}{ Package: \tab BayesTraitsAnalysis\cr
              Type: \tab Package\cr
              Version: \tab 0.1\cr
              Date: \tab 2016-06-17\cr
              License: \tab GPL\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{package}

