% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_BayesTraits_log.R
\name{read_BayesTraits_log}
\alias{read_BayesTraits_log}
\title{Read BayesTraits log}
\usage{
read_BayesTraits_log(x)
}
\arguments{
\item{x}{Log file to read}
}
\value{
`data.frame` with the contents of the log file.
}
\description{
Read in an arbitrary BayesTraits log file and return with
sanitized headers
}

